+++ .github/workflows/pr.yml
@@ -15,5 +15,5 @@ jobs:
         uses: srggrs/assign-one-project-github-action@1.2.1
         if: github.event.action == 'opened'
         with:
-          project: 'https://github.com/mysterix5/capstone/projects/1'
+          project: 'https://github.com/mysterix5/vover/projects/1'
           column_name: 'Review in progress'
