+++ Procfile
@@ -1,1 +1,1 @@
-web: java -XX:+UseContainerSupport -jar backend/target/capstone-backend-0.0.1-SNAPSHOT.jar
+web: java -XX:+UseContainerSupport -jar backend/target/vover-backend-0.0.1-SNAPSHOT.jar
