+++ backend/src/main/java/com/github/mysterix5/vover/model/WordTag.java
@@ -1,4 +1,4 @@
-package com.github.mysterix5.capstone.model;
+package com.github.mysterix5.vover.model;
 
 public enum WordTag {
     NORMAL, SAD, ACCENT
