+++ backend/src/main/java/com/github/mysterix5/vover/textHandling/TextController.java
@@ -1,8 +1,8 @@
-package com.github.mysterix5.capstone.textHandling;
+package com.github.mysterix5.vover.textHandling;
 
-import com.github.mysterix5.capstone.model.AudioResponseDTO;
-import com.github.mysterix5.capstone.model.WordResponseDTO;
-import com.github.mysterix5.capstone.model.TextSubmitDTO;
+import com.github.mysterix5.vover.model.AudioResponseDTO;
+import com.github.mysterix5.vover.model.WordResponseDTO;
+import com.github.mysterix5.vover.model.TextSubmitDTO;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
