+++ backend/src/main/java/com/github/mysterix5/vover/textHandling/TextService.java
@@ -1,10 +1,10 @@
-package com.github.mysterix5.capstone.textHandling;
+package com.github.mysterix5.vover.textHandling;
 
-import com.github.mysterix5.capstone.cloudstorage.CloudService;
-import com.github.mysterix5.capstone.model.AudioResponseDTO;
-import com.github.mysterix5.capstone.model.Availability;
-import com.github.mysterix5.capstone.model.WordDbEntity;
-import com.github.mysterix5.capstone.model.WordResponseDTO;
+import com.github.mysterix5.vover.cloudstorage.CloudService;
+import com.github.mysterix5.vover.model.AudioResponseDTO;
+import com.github.mysterix5.vover.model.Availability;
+import com.github.mysterix5.vover.model.WordDbEntity;
+import com.github.mysterix5.vover.model.WordResponseDTO;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
