+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "capstone-frontend",
+  "name": "vover-frontend",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
